Split statvars.py -> config.py and const.py
