+++ tickets_plus/database/__init__.py
@@ -1,14 +1,16 @@
 """General database package for Tickets+.
 
-This package contains the database layer, models and statvars.
+This package contains the database layer, models, const and config.
 It is used by the bot to store and retrieve data from the database.
-Or to get configuration values from the statvars module.
+Get configuration values from the confg module.
+And get constants from the const module.
 
 Typical usage example:
     ```py
     from tickets_plus.database import layer
     from tickets_plus.database import models
-    from tickets_plus.database import statvars
+    from tickets_plus.database import const
+    from tickets_plus.database import config
     ...
     ```
 """
